requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
albumentations==0.5.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.0
certifi==2021.5.30
chardet==3.0.4
click==8.1.7
colorama==0.4.6
configparser==5.0.1
cycler==0.10.0
decorator==4.4.2
dnspython==2.6.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.0.12
fsspec==2024.6.1
ftfy==6.2.0
future==0.18.2
gitdb==4.0.5
gitpython==3.1.12
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==2.10
imagehash==4.3.1
imageio==2.9.0
imgaug==0.4.0
jinja2==3.1.4
joblib==1.0.0
kaggle==1.5.10
kiwisolver==1.3.1
kornia==0.7.3
kornia-rs==0.1.5
markdown==3.3.3
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==5.1.0
munch==2.5.0
networkx==2.5
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.5
open-clip-torch==2.23.0
opencv-python==4.5.5.62
opencv-python-headless==4.1.2.30
orjson==3.10.6
packaging==24.1
pandas==1.2.0
pathtools==0.1.2
pillow==8.1.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.8.2
pydantic-core==2.20.1
pydicom==2.1.2
pygments==2.18.0
pyparsing==2.4.7
python==3.8.5
python-dateutil==2.8.1
python-dotenv==1.0.1
python-gdcm==3.0.24.1
python-multipart==0.0.9
python-slugify==4.0.1
pytorch-lightning==1.1.4
pytz==2020.5
pywavelets==1.1.1
pyyaml==5.3.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rich==13.7.1
rsa==4.7
sacremoses==0.0.43
safetensors==0.4.3
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.10.1
segmentation-models-pytorch==0.2.0
sentencepiece==0.2.0
sentry-sdk==0.19.5
setuptools==69.5.1
shapely==1.7.1
shellingham==1.5.4
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
sniffio==1.3.1
starlette==0.37.2
subprocess32==3.5.4
sympy==1.13.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2021.1.14
timm==1.0.8.dev0
tokenizers==0.19.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2
tqdm==4.56.0
transformers==4.42.4
typer==0.12.3
typing-extensions==4.12.2
ujson==5.10.0
urllib3==1.26.2
uvicorn==0.30.1
wandb==0.10.14
watchdog==0.10.4
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
werkzeug==1.0.1
wheel==0.43.0
xlrd==1.2.0
yarl==1.6.3
